requirements.txt
# Usar Python 3.11

# ---- Nucleo LangChain (usado no projeto hoje) ----
langchain-core
langchain-classic
langchain-openai

# ---- Integracoes opcionais para outros modelos (via LangChain) ----
langchain-ollama
langchain-google-genai
langchain-xai

# ---- RAG (backends) ----
langchain-weaviate
langchain-text-splitters
weaviate-client
ragie

# ---- Web Search ----
tavily-python
flashrank

# ---- Dependencias de runtime ----
tiktoken

# ---- Ferramentas de dados usadas ----
pandas
toons

# ---- Interfaces / APIs ----
flask
flask-cors
gunicorn

# ---- Configuração local ----
python-dotenv

# ---- Testes ----
pytest
pytest-cov
pytest-mock